refactor(PersonPod): separate chooser pod into its own file
